Graph.filter_edges_to returns points in the wrong order
Graph.filter_edges_to takes a point and returns all edges that terminate
at that point. For example, a->b terminates at point b, so
filter_edges_to will return it, and not, say, some edge that terminates
at a.

Currently filter_edges_to correctly finds those edges, but reverses
their points. So a->b would be reported as b->a.

This commit resolves that issue.